spawn_object supports children objects
